app/requirements.txt
fastapi
uvicorn
python-multipart
face_recognition
mysql-connector-python
python-dotenv